add .calculateIsoelectricPoint
Adds .calculateIsoelectricPoint to calculate the isoelectric point of peptides.
The general method is taken from:
    Moore, Dexter S.
    "Amino acid and peptide net charges: a simple calculational procedure."
    Biochemical Education 13.1 (1985): 10-11.
    http://dx.doi.org/10.1016/0307-4412(85)90114-1

Currently the pKa/charge values are taken from:
    Sillero, Antonio, and Andres Maldonado.
    "Isoelectric point determination of proteins and other macromolecules:
    oscillating method."
    Computers in biology and medicine 36.2 (2006): 157-166.
    http://dx.doi.org/10.1016/j.compbiomed.2004.09.006

Instead of a bisect search or an `optimize` approach to find the approximation
of the pI this method use a matrix based approach. All possible pHs are
calculated and the closest to one is taken. This has the disadvantage that
ridiculous precision for pH is not possible.